tint: fail on creation of a non-finite constant
This is a catch-all that handles all cases where an operation can result
in non-finite values, such as from calls to std::cosh and std::sinh.

Bug: tint:1581
Bug: tint:1747
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/110726
Kokoro: Kokoro <noreply+kokoro@google.com>

+++ src/dawn/tests/unittests/validation/OverridableConstantsValidationTests.cpp
@@ -247,7 +247,7 @@ TEST_F(ComputePipelineOverridableConstantsValidationTest, InvalidValue) {
     {
         // Valid:: Max
         std::vector<wgpu::ConstantEntry> constants{
-            {nullptr, "c3", std::numeric_limits<double>::max()}};
+            {nullptr, "c3", std::numeric_limits<float>::max()}};
         TestCreatePipeline(constants);
     }
 }
